fix: change image to azure linux (#15)
This pull request makes a small but important change to the
`Dockerfile`, updating the base image used in the run stage from a
distroless variant to the standard `aspnet:9.0-azurelinux3.0` image.
This may impact the runtime environment and available tools in the
container.

* Changed the run stage base image in `Dockerfile` from
`aspnet:9.0-azurelinux3.0-distroless` to `aspnet:9.0-azurelinux3.0`,
potentially affecting the runtime environment and available utilities.